Frontoffice/ContaCapitalFrontofficeEJB/src/test/java/br/com/sicoob/cca/frontoffice/negocio/servicos/interceptors/ContaCapitalFrontofficeInternacionalizacaoInterceptorTest.java
package br.com.sicoob.cca.frontoffice.negocio.servicos.interceptors;

import org.junit.Assert;
import org.junit.BeforeClass;
import org.junit.Test;

import br.com.bancoob.negocio.contexto.InformacoesUsuario;

public class ContaCapitalFrontofficeInternacionalizacaoInterceptorTest {

	@BeforeClass
	public static void setUpClass() {
		InformacoesUsuario.INSTANCIA.set(new InformacoesUsuario());
	}
	
	@Test
	public void deveInstanciarInterceptor() {
		ContaCapitalFrontofficeInternacionalizacaoInterceptor interceptor = 
				new ContaCapitalFrontofficeInternacionalizacaoInterceptor();
		Assert.assertNotNull(interceptor);
		Assert.assertNotNull(interceptor.getResourceBundle());
	}
	
}
